backend/requirements.txt
# App
Django==4.1.13
django-filter==22.1
djangorestframework==3.12.2
Markdown==3.3
python-decouple==3.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
django-test-without-migrations==0.6
django_nose==1.4.7
model_bakery==1.10.1
django-cors-headers==3.7.0
django_extensions==3.1.1
boto3==1.26.35
django-storages==1.11.1
pytest==7.1.3
pytest-django==4.1.0
pytest-mock==3.5.1
djangorestframework-simplejwt==4.6.0
Faker==8.0.0
requests==2.25.1
drf-yasg==1.20.0
sentry-sdk==1.0.0
drf-nested-routers==0.93.4
django-multiselectfield==0.1.12
ipython==8.15.0
pytz==2024.2
redis==5.2.1
celery==5.4.0
dj-database-url==0.5.0
dj-static==0.0.6
amazon-textract-textractor==1.8.5
ruff==0.9.7
pytest-vcr==1.0.2
vcrpy==7.0.0